requirements.txt
click==6.6
requests==2.9.1
PyYAML==3.11
